Extract generic ring_switch_tests (temporary solution)
