+++ crates/compute_test_utils/src/lib.rs
@@ -3,3 +3,4 @@
 pub mod bivariate_sumcheck;
 pub mod layer;
 pub mod piop;
+pub mod ring_switch;
